Improved the estimation of the remaining time during file transfers
This estimation adds an other variable: the initialisation (and finalisation) time needed for the transfer of a single file.
In the case of small files, this time becomes more important than the transfer rate.

The time for a single file is assumed to be t1 = t0 + S w0,
where t1 is the total time for the transfer of the file,
      t0 is the initialisation time,
      S is the size of the file,
      w0 is the inverse of the (internal) transfer rate.
Then sample variances the total transfer time and size are computed, and t0 and w0 can be deduced from these statistics.

See docs/algo_estimation_du_temps_de_transfer_de_fichiers.md for further details (in French).